Prepare monorepo import
