--- .eslintrc.json
@@ -1,24 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "es6": true,
-    "node": true
-  },
-  "extends": [
-    "eslint:recommended",
-    "plugin:vue/recommended",
-    "plugin:prettier/recommended",
-    "prettier",
-    "plugin:jest/recommended"
-  ],
-  "parserOptions": {
-    "parser": "@babel/eslint-parser",
-    "sourceType": "module"
-  },
-  "rules": {
-    "require-await": "warn"
-  },
-  "plugins": [
-    "jest"
-  ]
-}
