--- .github/config.yml
@@ -1,13 +1,0 @@
-# Configuration for update-docs - https://github.com/behaviorbot/update-docs
-
-# Comment to be posted to on PRs that don't update documentation
-updateDocsComment: >
-  Thanks for opening this pull request! The maintainers of this repository would appreciate it if you would create a [changelog](https://github.com/owncloud/owncloud-design-system/blob/master/changelog/README.md) item based on your changes.
-
-updateDocsWhiteList:
-  - Tests-only
-  - tests-only
-  - Tests-Only
-
-updateDocsTargetFiles:
-  - changelog/unreleased/
