--- .github/issue_template.md
@@ -1,34 +1,0 @@
-<!--
-Thanks for submitting a change to ownCloud!
-
-This is the issue tracker for the design system. 
-
-For reporting potential security issues please see https://owncloud.org/security/
-
-To make it possible for us to help you please fill out below information carefully.
-
-Before reporting any issues please make sure that you're using the latest available version of the ODS
---> 
-
-### Steps to reproduce
-1.
-2.
-3.
-
-### Expected behaviour
-Tell us what should happen
-
-### Actual behaviour
-Tell us what happens instead
-
-### Screenshots (if appropriate):
-
-
-#### Browser log
-```
-Insert your browser log here, this could for example include:
-
-a) The javascript console log
-b) The network log 
-c) ...
-```
