--- .github/pull_request_template.md
@@ -1,61 +1,0 @@
-<!--
-Thanks for submitting a change to ownCloud!
-
-This is the issue tracker for the design system. 
-
-For fixing potential security issues please see https://owncloud.org/security/
-
-To make it possible for us to get your change reviewed and merged please carefully fill out the requested information below.
-
-Please note that any kind of change needs first be submitted to the master branch which holds the next major version of the ownCloud design system.
-
-We will carefully discuss if your change can or has to be backported to stable branches.
-
-Please set the following labels:
-
-- Set label "3 - To review" for review or "2 - Development" in case the PR still has open tasks
-- Assignment: assign to self
--->
-
-## Description
-<!--- Describe your changes in detail -->
-
-## Related Issue
-<!--- This project only accepts pull requests related to open issues -->
-<!--- If suggesting a new feature or change, please discuss it in an issue first -->
-<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
-<!--- Please link to the issue here: -->
-- Fixes <issue_link>
-
-## Motivation and Context
-<!--- Why is this change required? What problem does it solve? -->
-
-## How Has This Been Tested?
-<!--- Please describe in detail how you tested your changes. -->
-<!--- Include details of your testing environment, and the tests you ran to -->
-<!--- see how your change affects other areas of the code, etc. -->
-- test environment:
-- test case 1:
-- test case 2:
-- ...
-
-## Screenshots (if appropriate):
-
-## Types of changes
-<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
-- [ ] Bug fix (non-breaking change which fixes an issue)
-- [ ] New feature (non-breaking change which adds functionality)
-- [ ] Breaking change (fix or feature that would cause existing functionality to change)
-- [ ] Technical debt
-- [ ] Tests
-
-## Checklist:
-<!-- Tick the checkboxes when done. -->
-- [ ] Code changes
-- [ ] Unit tests added
-- [ ] Acceptance tests added
-- [ ] Documentation added/updated
-
-## Open tasks:
-<!-- In case of incomplete PR, please list the open tasks here -->
-- [ ] ...
