+++ .gitignore
@@ -1,25 +1,2 @@
-.DS_Store
-.eslintcache
-node_modules/
-dist/
-coverage/
-.codacy-coverage/
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.stylelintcache
-.pnpm-store/
-
-# Editor directories and files
-.idea
-.vscode
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-
 # Ignore compiled tokens
 src/assets/tokens/
-
-# Ignore SonarCloud output
-.scannerwork
