--- .npmrc
@@ -1,4 +1,0 @@
-auto-install-peers=true
-hoist-pattern[]=*babel*
-hoist-pattern[]=postcss-scss
-hoist-pattern[]=convert-source-map
