--- .prettierignore
@@ -1,4 +1,0 @@
-package.json
-package-lock.json
-src/assets/tokens/*
-dist/*
