--- .prettierrc
@@ -1,22 +1,0 @@
-{
-  "printWidth": 100,
-  "singleQuote": false,
-  "trailingComma": "es5",
-  "bracketSpacing": true,
-  "jsxBracketSameLine": false,
-  "semi": false,
-  "requirePragma": false,
-  "insertPragma": false,
-  "useTabs": false,
-  "tabWidth": 2,
-  "arrowParens": "avoid",
-  "proseWrap": "preserve",
-  "overrides": [
-    {
-      "files": "*.scss",
-      "options": {
-        "parser": "scss"
-      }
-    }
-  ]
-}
