+++ docs/components/tokens/ColorTokens.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script>
-import designTokens from "@/assets/tokens/ods.json"
+import designTokens from '../../../src/assets/tokens/ods.json'
 
 /**
  * The color palette comes with 5 different weights for each hue. These hues
