+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "owncloud-design-system",
-  "version": "14.0.0-alpha.25",
+  "name": "@ownclouders/design-system",
+  "version": "15.0.0-alpha.1",
   "description": "ownCloud Design System is based on VueDesign Systems and is used to design ownCloud UI components",
   "keywords": [
     "vue design system",
@@ -13,20 +13,15 @@
   "homepage": "https://owncloud.design/",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/owncloud/owncloud-design-system.git"
+    "url": "git+https://github.com/owncloud/web.git"
   },
   "license": "AGPL-3.0",
   "author": "ownClouders",
-  "main": "dist/system/system.js",
-  "files": [
-    "dist/system"
-  ],
+  "main": "src/index.ts",
   "scripts": {
     "start": "npm-run-all --parallel styleguide tokens:w",
-    "build:system": "npm-run-all tokens node:build:system",
     "build:docs": "npm-run-all tokens styleguide:build",
-    "lint": "npm-run-all --parallel lint:eslint lint:stylelint",
-    "lint:eslint": "eslint '**/*.{js,vue}' --cache",
+    "lint": "npm-run-all --parallel lint:stylelint",
     "lint:stylelint": "stylelint './src/styles/**/*.scss' --cache",
     "node:build:system": "node ./build/build-system.js",
     "styleguide": "vue-styleguidist server --open --config ./config/docs.config.js",
@@ -46,9 +41,8 @@
     "not OperaMobile > 0"
   ],
   "devDependencies": {
-    "@adobe/leonardo-contrast-colors": "^1.0.0-alpha.8",
+    "@adobe/leonardo-contrast-colors": "1.0.0-alpha.8",
     "@babel/core": "^7.16.5",
-    "@babel/eslint-parser": "^7.16.5",
     "@babel/plugin-proposal-class-properties": "^7.16.5",
     "@babel/plugin-proposal-object-rest-spread": "^7.16.5",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
@@ -69,12 +63,6 @@
     "css-loader": "^5.0.1",
     "deepmerge": "^4.2.2",
     "depcheck": "^1.3.1",
-    "easygettext": "^2.16.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jest": "^26.0.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^8.5.0",
-    "eslint": "^8.19.0",
     "file-loader": "^6.2.0",
     "filesize": "^9.0.9",
     "focus-trap-vue": "^1.1.1",
@@ -126,8 +114,7 @@
     "vue-select": "^3.12.0",
     "vue-style-loader": "^4.1.2",
     "vue-styleguidist": "^4.44.2",
-    "vue-template-compiler": "2.7.13",
-    "vue": "2.7.13",
+    "vue": "2.7.14",
     "webfontloader": "^1.6.28",
     "webpack-bundle-analyzer": "^4.2.0",
     "webpack-merge-and-include-globally": "^2.1.24",
