--- sonar-project.properties
@@ -1,35 +1,0 @@
-
-sonar.projectKey=owncloud_owncloud-design-system
-sonar.organization=owncloud-1
-sonar.projectName=owncloud-design-system
-sonar.projectVersion=14.0.0-alpha.25
-sonar.host.url=https://sonarcloud.io
-
-# =====================================================
-#   Meta-data for the project
-# =====================================================
-
-sonar.links.homepage=https://github.com/owncloud/owncloud-design-system
-sonar.links.ci=https://drone.owncloud.com/owncloud/owncloud-design-system/
-sonar.links.scm=https://github.com/owncloud/owncloud-design-system
-sonar.links.issue=https://github.com/owncloud/owncloud-design-system/issues
-
-# =====================================================
-#   Properties that will be shared amongst all modules
-# =====================================================
-
-# SQ standard properties
-sonar.sources=.
-
-# Pull Requests
-sonar.pullrequest.provider=github
-sonar.pullrequest.github.repository=owncloud/owncloud-design-system
-sonar.pullrequest.base=${env.SONAR_PULL_REQUEST_BASE}
-sonar.pullrequest.branch=${env.SONAR_PULL_REQUEST_BRANCH}
-sonar.pullrequest.key=${env.SONAR_PULL_REQUEST_KEY}
-
-# Properties specific to language plugins:
-sonar.javascript.lcov.reportPaths=coverage/lcov.info
-
-# Exclude files
-sonar.exclusions=build/**,docs/**,node_modules/**,**/*.spec.js,dist/**,changelog/**,config/**,release/**,package.json,jest.conf.js,jest.setup.js,Makefile,CHANGELOG.md,README.md
