+++ src/components/atoms/OcButton/OcButton.vue
@@ -176,6 +176,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 @mixin oc-button-gap($factor) {
   gap: round($oc-space-small * $factor / 2) * 2;
 }
