+++ src/components/atoms/OcCheckbox/OcCheckbox.vue
@@ -139,6 +139,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 @mixin oc-form-check-size($factor) {
   height: $oc-size-form-check-default * $factor;
   width: $oc-size-form-check-default * $factor;
