+++ src/components/atoms/OcGrid/OcGrid.vue
@@ -49,6 +49,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 // Ported from UiKit until we change UI frameworks
 
 // Name:            Grid
