+++ src/components/atoms/OcProgress/OcProgress.vue
@@ -91,6 +91,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 $progress-height: 15px !default;
 $progress-height-small: 5px !default;
