+++ src/components/atoms/OcSearchBar/OcSearchBar.vue
@@ -284,6 +284,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 .oc-search {
   @extend .oc-flex-middle;
