+++ src/components/atoms/OcSelect/OcSelect.vue
@@ -134,6 +134,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 .vs--disabled {
   .vs__clear,
   .vs__dropdown-toggle,
