+++ src/components/atoms/OcTableSimple/OcTableSimple.vue
@@ -32,6 +32,8 @@ export default {
 }
 </script>
 <style lang="scss">
+@import '../../../styles/styles';
+
 .oc-table-simple {
   border-collapse: collapse;
   border-spacing: 0;
