+++ src/components/atoms/OcTag/OcTag.vue
@@ -85,6 +85,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 .oc-tag {
   align-items: center;
   background-color: var(--oc-color-background-default);
