+++ src/components/atoms/OcTextarea/OcTextarea.vue
@@ -189,6 +189,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 .oc-textarea {
   background-color: var(--oc-color-input-bg);
   border: 1px solid var(--oc-color-input-border);
