+++ src/components/molecules/OcTable/OcTable.vue
@@ -466,6 +466,8 @@ export default {
 }
 </script>
 <style lang="scss">
+@import '../../../styles/styles';
+
 .oc-table {
   border-collapse: collapse;
   border-spacing: 0;
