+++ src/components/molecules/index.ts
@@ -1,0 +1,16 @@
+
+export { default as OcAccordionItem } from './OcAccordionItem/OcAccordionItem.vue'
+export { default as OcAvatar } from './OcAvatar/OcAvatar.vue'
+export { default as OcAvatarFederated } from './OcAvatarFederated/OcAvatarFederated.vue'
+export { default as OcAvatarGroup } from './OcAvatarGroup/OcAvatarGroup.vue'
+export { default as OcAvatarGuest } from './OcAvatarGuest/OcAvatarGuest.vue'
+export { default as OcAvatarLink } from './OcAvatarLink/OcAvatarLink.vue'
+export { default as OcBreadcrumb } from './OcBreadcrumb/OcBreadcrumb.vue'
+export { default as OcContextualHelper } from './OcContextualHelper/OcContextualHelper.vue'
+export { default as OcInfoDrop } from './OcInfoDrop/OcInfoDrop.vue'
+export { default as OcLogo } from './OcLogo/OcLogo.vue'
+export { default as OcPageSize } from './OcPageSize/OcPageSize.vue'
+export { default as OcPagination } from './OcPagination/OcPagination.vue'
+export { default as OcStatusIndicators } from './OcStatusIndicators/OcStatusIndicators.vue'
+export { default as OcTable } from './OcTable/OcTable.vue'
+export { default as OcTextInput } from './OcTextInput/OcTextInput.vue'
