+++ src/components/organisms/OcAccordion/OcAccordion.vue
@@ -148,6 +148,8 @@ export default {
 </script>
 
 <style lang="scss">
+@import '../../../styles/styles';
+
 .oc-accordion {
   padding: 0;
