+++ src/directives/index.ts
@@ -1,0 +1,1 @@
+export { default as OcTooltip } from './OcTooltip'
