+++ src/styles/styles.scss
@@ -1,4 +1,5 @@
 // 1. Your custom variables and variable overwrites.
+@import "../assets/tokens/ods";
 @import "theme/variables";
 
 // 2. Import Vendor.
